refactor : refactor on Client
refactoring slices and apis
