+++ projects/client/src/apps/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
-import AdminReducer from "../feature/admin/AdminSlice";
-import AdminLogInReducer from "../feature/admin/AdminLogInSlice";
+import AdminReducer from "../feature/admin/slice/AdminSlice";
+import AdminLogInReducer from "../feature/admin_auth/slice/AdminLogInSlice";
 
 export default configureStore({
   reducer: {
