+++ projects/client/src/components/admin/manage_user/ManageAdmin.jsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { getAllAdmin, getSingleWarehouseAdmin, getWarehouseCities } from "../../../feature/admin/AdminSlice";
 import RenderAdminsData from "./all_admin/RenderAdminsData";
 import Pagination from "./all_admin/Pagination";
 import AddNewAdmin from "./all_admin/create_data/AddNewAdmin";
 import DeleteAdminModal from "./all_admin/delete_data/DeleteAdminModal";
 import EditAdminModal from "./all_admin/edit_data/EditAdminModal";
+import { getAllAdmin, getSingleWarehouseAdmin, getWarehouseCities } from "../../../feature/admin";
 
 function ManageAdmin(props) {
   let { page } = props;
