+++ projects/client/src/components/admin/manage_user/all_admin/RenderAdminsData.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useDispatch } from "react-redux";
-import { getSingleUser } from "../../../../feature/admin/AdminSlice";
+import { getSingleUser } from "../../../../feature/admin";
 
 function RenderAdminsData(props) {
   const { allAdmin, editBtnHndler, setDeleteClicked } = props;
