+++ projects/client/src/feature/admin/api/deleteUserData.js
@@ -1,0 +1,17 @@
+import { axiosInstance } from "../slice/AdminSlice";
+
+export const deleteUserData = async (id_user) => {
+  try {
+    const TOKEN = localStorage.getItem("admin_token");
+    const response = await axiosInstance.patch(
+      `/user/${id_user}`,
+      {},
+      {
+        headers: { Authorization: `Bearer ${TOKEN}` },
+      },
+    );
+    return response.data;
+  } catch (error) {
+    alert(error.response.data.message);
+  }
+};
