+++ projects/client/src/feature/admin/api/updateAdminWarehouse.js
@@ -1,0 +1,17 @@
+import { axiosInstance } from "../slice/AdminSlice";
+
+export const updateAdminWarehouse = async (data) => {
+  try {
+    const TOKEN = localStorage.getItem("admin_token");
+    const response = await axiosInstance.patch(
+      "/",
+      { ...data },
+      {
+        headers: { Authorization: `Bearer ${TOKEN}` },
+      },
+    );
+    return response.data;
+  } catch (error) {
+    alert(error.response.data.message);
+  }
+};
