+++ projects/client/src/feature/admin/index.js
@@ -1,0 +1,9 @@
+export * from "./api/getAllUserData";
+export * from "./api/getSingleUser";
+export * from "./api/getAllAdmin";
+export * from "./api/getSingleWarehouseAdmin";
+export * from "./api/getWarehousesCities";
+export * from "./api/getWarehousesInCities";
+export * from "./api/updateAdminWarehouse";
+export * from "./api/deleteUserData";
+export * from "./api/createNewAdmin";
