+++ projects/client/src/feature/admin_warehouse/index.js
@@ -1,0 +1,6 @@
+export * from "./api/getWarehouses";
+export * from "./api/deleteWarehouse";
+export * from "./api/getProvinces";
+export * from "./api/getCitiesByProvinces";
+export * from "./api/createNewWarehouse";
+export * from "./api/editWarehouse";
