+++ projects/client/src/feature/admin_warehouse/slice/AdminWarehouseSlice.js
@@ -1,0 +1,3 @@
+import axios from "axios";
+const API_ADMIN_WAREHOUSE = process.env.REACT_APP_ADMIN_WAREHOUSE_URL;
+export const axiosInstance = axios.create({ baseURL: API_ADMIN_WAREHOUSE });
