+++ projects/client/src/pages/UserManagement.jsx
@@ -3,7 +3,7 @@ import LayoutAdmin from "../components/LayoutAdmin";
 import UserMgtUserData from "../components/admin/manage_user/UserMgtUserData";
 import ManageAdmin from "../components/admin/manage_user/ManageAdmin";
 import { useDispatch } from "react-redux";
-import { getAllAdmin } from "../feature/admin/AdminSlice";
+import { getAllAdmin } from "../feature/admin";
 
 function UserManagement() {
   const [allDataBtnClicked, setAllDataBtnClicked] = useState(true);
