+++ projects/client/src/pages/WarehouseMgt.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import LayoutAdmin from "../components/LayoutAdmin";
 import HeaderWarehouseMgt from "../components/admin/manage_warehouse/HeaderWarehouseMgt";
 import Body from "../components/admin/manage_warehouse/Body";
-import { getWarehouses } from "../feature/admin/AdminWarehouseSlice";
+import { getWarehouses } from "../feature/admin_warehouse";
 
 function WarehouseMgt() {
   const [warehouses, setWarehouses] = useState([]);
